requirements.txt
Adafruit-Blinka==0.2.7
adafruit-circuitpython-busdevice==2.2.5
adafruit-circuitpython-vl6180x==1.1.0
Adafruit-GPIO==1.0.3
Adafruit-PureIO==0.2.3
atomicwrites==1.2.1
attrs==18.2.0
bcrypt==3.1.4
certifi==2018.10.15
cffi==1.11.5
chardet==3.0.4
click==6.7
dominate==2.3.4
Flask==1.0.2
Flask-Bcrypt==0.7.1
Flask-Bootstrap4==4.0.2
Flask-Cors==3.0.6
Flask-Login==0.4.1
Flask-WTF==0.14.2
idna==2.7
itsdangerous==0.24
Jinja2==2.10
MarkupSafe==1.0
more-itertools==4.3.0
pathlib2==2.3.2
pkg-resources==0.0.0
pluggy==0.7.1
py==1.6.0
pycparser==2.19
pytest==3.8.2
pytest-dotenv==0.2.0
pytest-flask==0.13.0
python-dotenv==0.9.1
requests==2.20.0
rpi-ws281x==4.0.0
RPi.GPIO==0.6.3
six==1.11.0
spidev==3.2
urllib3==1.24
visitor==0.1.3
Werkzeug==0.14.1
WTForms==2.2.1
